requirements.txt
captcha
tensorflow
Pillow
numpy

